fix: issue when invalid value for env is set
In case of negative value for the number of threads or the array length
being smaller than the number of threads set will cause either an
invalid use case or unnecessary resource utilisation. So in case of
value of array less than the threads give preference to array length.

In case of an invalid env number default fallback to run in a single
thread.