fix: chunk allocation to threads
When the array is getting chunked the same array was being delegated to
each thread which was cause race conditions while writing.

Delegated each index to a single thread and fixed the chunking logic to
create chunks by diving the length of array to number of threads and
taking ceiling of that it might not be absolutely divisible.